prototypes

+++ libft/srcs/substr.c
@@ -1,6 +1,6 @@
-#include <stdlib.h>
+#include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, unsigned int len)
+char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*sub;
 	unsigned int	i;
